architect with blueprint,rewrite sql syntex with db object, loading effect
