+++ taipei_day_trip_project/static/css/thankyou/thankyou.css
