+++ taipei_day_trip_project/static/image/icon/icon_delete.png
