+++ taipei_day_trip_project/static/javascript/attraction/attraction.js
