+++ taipei_day_trip_project/static/javascript/boking/booking.js
