+++ taipei_day_trip_project/static/javascript/index/index_login_signup.js
