+++ taipei_day_trip_project/static/javascript/thankyou/thankyou.js
