+++ taipei_day_trip_project/templates/booking.html
