+++ taipei_day_trip_project/templates/index.html
@@ -97,9 +97,9 @@
         <!-- picture -->
         <div class="p0">
             <div class="p1"></div>
+            <!-- loading gif before api finished -->
+            <div class="loading"></div>
         </div>
-        <!-- intersct observe item -->
-        <!-- <div class="obt"></div> -->
         <!-- last -->
         <div class="l1">
             <div class="l2">COPYRIGHT © 2021 台北一日遊</div>
