--- test.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Document</title>
-        <link rel="stylesheet" href="/test.css" />
-    </head>
-    <body>
-        <div class="b">
-            <button>案案案ㄢ</button>
-        </div>
-        <div class="a">
-            <button>aaaaaaa</button>
-        </div>
-    </body>
-    <script src="/test.js"></script>
-</html>
